trunk_food vs trunk: only aconfig flags differ
Change trunk_food.scl to be a minimal difference from trunk, with a note
saying that it should not be changed.  This will keep the build flags
for trunk and trunk_food in sync.

Bug: b/313479515
Test: manual
Ignore-AOSP-First: Corresponding changeset is in AOSP
Merged-In: I28092379c7a5a73a741b6f54e079c05973736672
Merged-In: Idef2c06b341a105cb95ee70fc0bc4997229bd787
Change-Id: Idef2c06b341a105cb95ee70fc0bc4997229bd787